feat(shake-it): ✨ add nav menu and food recommendation route
