+++ apps/shake-it/src/app/_layout.tsx
@@ -37,7 +37,7 @@ export default function RootLayout() {
       <ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
         <Stack>
           <Stack.Screen name="index" options={{ headerShown: false }} />
-          <Stack.Screen name="explore" options={{ headerShown: false }} />
+          <Stack.Screen name="food" options={{ title: "오늘 뭐 먹지?" }} />
           <Stack.Screen name="+not-found" />
         </Stack>
       </ThemeProvider>
